Add Android Instrumented Test Environment
This commit adds android instrumented test environment.
We need to test and know whether caver-java-android lib
is well-working on real android machine or not.

For this
1 > Add android_instrumented_test sub-project
This sub-project is made for instrumented testing.
This sub-project will use unit test files already written at core.

This sub-project uses AndroidX Test Library for instrumented test.

2 > Add Suite Test for TestClasses containing nested classes
This makes AndroidX Test Library well works on TestClass which contains nested classes.
Without it, test cases of nested classes does not work on emulator.

3 > Update circle-ci config and add shell script
To re-use unit tests written at core, I added a script for it and updated config of circle-ci too.

4 > Update git ignore
To make android_instrumented_test which is Android Project works well, we must not ignore
`gradle.properties` because it contains essentials of AndroidX Test Library.