Update Default Gas Provider
DefaultGasProvider now implements ContractGasProvider instead of extends StaticGasProvider to provide dynamic gas price value based on network.
Because Klaytn is now planning to apply dynamic gas price policy.

If that policy is activated and we does not change this behavior,
there is a high probability that the transactions created using the SDK cannot be processed.